restructuing beforeEntry
